TDD - unit tests written
